Temporarily renamed House.py again due to a weird merge conflict
