+++ GameLib/TempHouse.py
